+++ dash_chat/_imports_.py
@@ -1,5 +1,3 @@
 from .ChatComponent import ChatComponent
 
-__all__ = [
-    "ChatComponent"
-]
+__all__ = ["ChatComponent"]