+++ src/styles/sharedStyles.js
@@ -1,7 +1,0 @@
-export const bubbleStyle = {
-    maxWidth: "75%",
-    marginBottom: "10px",
-    padding: "10px",
-    borderRadius: "12px",
-    wordWrap: "break-word"
-};